BUG: Fix mouse cursor not being changed for views not mapped in layout
When changing mouse modes, the cursor is not updated for views that are not mapped in the current layout. This causes an incorrect cursor to be displayed in views that are not part of the layout when the mouse mode is changed.

Example:
- Start in four-up layout
- Change to window/level mouse mode
- Maximize red slice view
- Change to view transform mouse mode
- Restore view layout
- Mouse cursor still shows window/level cursor in all views except the red slice == *Error*

Fixed by updating view cursors even when the view is not mapped in the current layout.